Close manifest-PR upon closing PR
When closing a PR it doesn't make sense to keep the manifest-PR
open.
Do this automatically to avoid that we forget doing this.

To make it clear why the PR was closed, it also adds a closing
comment.

We delete the branch as well as we assume there are no changes
there that we want to keep.

Ignore the error code in case the PR didn't exist.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>